src/requirements.txt
configparser==5.0.1
paho-mqtt==1.5.1
